requirements.txt
streamlit
google-generativeai
pandas
openpyxl
altair
numpy
pillow
jinja2
jsonschema
narwhals
pyarrow
pydeck
python-dateutil
pytz
requests
requests-ntlm
tenacity
tornado
watchdog
